+++ crates/assembly-syntax/src/ast/visit.rs
@@ -476,7 +476,7 @@ where
         | MemStoreImm(imm)
         | MemStoreWBeImm(imm)
         | MemStoreWLeImm(imm) => visitor.visit_immediate_u32(imm),
-        EmitImm(imm) => visitor.visit_immediate_felt(imm),
+        EmitImm(imm) | TraceImm(imm) => visitor.visit_immediate_felt(imm),
         SysEvent(sys_event) => visitor.visit_system_event(Span::new(span, sys_event)),
         Exec(target) => visitor.visit_exec(target),
         Call(target) => visitor.visit_call(target),
@@ -505,7 +505,7 @@ where
         | MemStoreWLe | MemStream | AdvPipe | AdvPush | AdvPushW | AdvLoadW | Hash | HMerge
         | HPerm | MTreeGet | MTreeSet | MTreeMerge | MTreeVerify | FriExt2Fold4 | DynExec
         | DynCall | DebugVar(_) | HornerBase | HornerExt | CryptoStream | EvalCircuit
-        | LogDeferred | Emit => ControlFlow::Continue(()),
+        | LogDeferred | Emit | Trace => ControlFlow::Continue(()),
     }
 }
 
@@ -1040,7 +1040,7 @@ where
         | MemStoreImm(imm)
         | MemStoreWBeImm(imm)
         | MemStoreWLeImm(imm) => visitor.visit_mut_immediate_u32(imm),
-        EmitImm(imm) => visitor.visit_mut_immediate_felt(imm),
+        EmitImm(imm) | TraceImm(imm) => visitor.visit_mut_immediate_felt(imm),
         SysEvent(sys_event) => visitor.visit_mut_system_event(Span::new(span, sys_event)),
         Exec(target) => visitor.visit_mut_exec(target),
         Call(target) => visitor.visit_mut_call(target),
@@ -1069,7 +1069,7 @@ where
         | MemStoreWLe | MemStream | AdvPipe | AdvPush | AdvPushW | AdvLoadW | Hash | HMerge
         | HPerm | MTreeGet | MTreeSet | MTreeMerge | MTreeVerify | FriExt2Fold4 | DynExec
         | DynCall | DebugVar(_) | HornerBase | HornerExt | EvalCircuit | CryptoStream
-        | LogDeferred | Emit => ControlFlow::Continue(()),
+        | LogDeferred | Emit | Trace => ControlFlow::Continue(()),
     }
 }